refactored login api calls in server.js to be in a seperate file
