+++ server/server.js
@@ -4,15 +4,9 @@ const path = require('path');
 const mongoose = require('mongoose');
 const config = require('config');
 const cors = require('cors');
-const bodyParser = require('body-parser');
-const User = require("./models/User.js");
-const bcrypt = require("bcryptjs");
-const jwt = require("jsonwebtoken");
-const JWT_SECRET="jsdjfsjfksdfjhsdjfhdsfkjsdhf87879837937987*&&%^$%$^&^&^&^ksjhfkdhfksdhkfjhdskfjhdsk";
-const fs = require('fs');
-
 
 const gameHandlers = require('./handlers/game.js');
+const loginHandlers = require('./handlers/login.js')
 
 const app = express();
 const port = 8080;
@@ -33,105 +27,14 @@ mongoose.connect(db,//connect to db
 
 app.get('/getPrompt', gameHandlers.getPrompt); // Would change url
 
+app.post('/changePassword', loginHandlers.changePassword);
 
-app.post('/api/change-password', async (req, res) => {
-	const { token, newpassword: plainTextPassword } = req.body
+app.post('/validateLogin', loginHandlers.validateLogin);
 
-	if (!plainTextPassword || typeof plainTextPassword !== 'string') {
-		return res.json({ status: 'error', error: 'Invalid password' })
-	}
+app.post('/register', loginHandlers.register)
 
-	if (plainTextPassword.length < 5) {
-		return res.json({
-			status: 'error',
-			error: 'Password too small. Should be atleast 6 characters'
-		})
-	}
+//app.post('/addToken', loginHandlers.addToken);
 
-	try {
-		const user = jwt.verify(token, JWT_SECRET)
-
-		const _id = user.id
-
-		const password = await bcrypt.hash(plainTextPassword, 10)
-
-		await User.updateOne(
-			{ _id },
-			{
-				$set: { password }
-			}
-		)
-		res.json({ status: 'ok' })
-	} catch (error) {
-		console.log(error)
-		res.json({ status: 'error', error: ';))' })
-	}
-})
-
-app.post('/api/login', async (req, res) => {
-	const { username, password } = req.body
-	const user = await User.findOne({ username }).lean()
-
-	if (!user) {
-		return res.json({ status: 'error', error: 'Invalid username/password' })
-	}
-
-	if (await bcrypt.compare(password, user.password)) {
-		// the username, password combination is successful
-
-		const token = jwt.sign(
-			{
-				id: user._id,
-				username: user.username
-			},
-			JWT_SECRET
-		)
-
-		return res.json({ status: 'ok', data: token })
-	}
-
-	res.json({ status: 'error', error: 'Invalid username/password' })
-})
-
-app.post('/api/register', async (req, res)=>{
-	console.log(req.body);
-
-	const {username, password: plainTextPassword } = req.body
-	const user = await User.findOne({username}).lean();
-
-	if(!username || typeof username !== 'string'){
-		return res.json({status:"error", error:"Invalid username"})
-	}
-
-	if(!plainTextPassword || typeof plainTextPassword !== "string") {
-		return res.json({status:"error", error:"Invalid password"})
-	}
-	if (plainTextPassword.length < 4 ) {
-		return res.json({status:"error", error:"Password is too short"})
-	}
-	if(user){
-		return res.json({status: 'error', error: 'Username already exists'})
-	}
-	const password = await bcrypt.hash(plainTextPassword, 10)
-	try{
-		await User.create({
-			username,
-			password,
-			gamesPlayed:0
-		})
-	}catch(error) {
-		console.log(error)
-		return res.json({status:"error"})
-	}
-	const reguser = await User.findOne({username}).lean();
-	const token = jwt.sign(
-		{
-			id: reguser._id,
-			username: reguser.username
-		},
-		JWT_SECRET
-	)
-	return res.json({status:"ok", data:token});
-})
+//app.get('/validateToken/:token', loginHandlers.validateToken);
 
 app.listen(port, () => console.log(`Server listening on http://localhost:${port}`));
